Add nltk Update code to use lemmas and tokens from nltk Update list of known words
The reason to switch back to nltk was simple. Without it I need to write a
lot of boilerplate code and basically reinvent the wheel.
Tokenization solves an issue with separating words from the string.
Lemmatization gets the basic form of word, it ommits endings. It works very well
for majority of cases, however in some cases it fails.